created tests for important functions in the game score system
